docs: [1.X] fixes of PVP exceptions (#3300)
This PR focuses on fixing PVP errors related to PVP-151-1 "**_Public
APIs should be documented_**" and all remaining PVP errors **present in
pvpExceptions.json file**

We have quite a lot of errors to address so this PR will be a
collaborative effort. If anyone want's to add to it what we should do
is:

1. Check errors present in pvpExceptions.json file (preferably under
"PVP-150-1", choose some to fix
2. Address those errors and when making commit remove the fixed errors
from pvpExceptions file
The first commit of this PR is an example of how such process should
look like

In general [new CI
implementation](https://github.com/Unity-Technologies/com.unity.netcode.gameobjects/pull/3193)
will catch all new errors like this so all we need to do is fix the
present ones. This PR **does not** focus on PVP present in gold profile
but if capacity allows, those can also be addressed.

**Error present in "PVP-41-1" is to be expected** so there is no need of
fixing this one!

Another thing is that we could ignore error related to missing APIs for
RuntimeTests because those were made internal in 2.X version (that's why
those errors are not popping by there). Those APIs are not user facing
so we could just leave them (otherwise there is like 2000 more APIs to
add)

---------

Co-authored-by: Noel Stephens <noel.stephens@unity3d.com>
Co-authored-by: Emma <emma.mcmillan@unity3d.com>
Co-authored-by: unity-renovate[bot] <120015202+unity-renovate[bot]@users.noreply.github.com>